+++ de.bitub.step.p21/src/de/bitub/step/p21/XPressModelConstants.java
@@ -1,4 +1,4 @@
-package de.bitub.step.p21.util;
+package de.bitub.step.p21;
 
 public interface XPressModelConstants
 {
