+++ de.bitub.step.p21/src/de/bitub/step/p21/parser/P21EntityListener.java
@@ -1,4 +1,4 @@
-package de.bitub.step.p21;
+package de.bitub.step.p21.parser;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -12,6 +12,11 @@ import org.eclipse.emf.ecore.util.EcoreUtil;
 
 import com.google.inject.Inject;
 
+import de.bitub.step.p21.IndexUtil;
+import de.bitub.step.p21.P21Index;
+import de.bitub.step.p21.StepParser;
+import de.bitub.step.p21.StepUntypedToEcore;
+import de.bitub.step.p21.XPressModel;
 import de.bitub.step.p21.StepParser.IntegerContext;
 import de.bitub.step.p21.StepParser.ListContext;
 import de.bitub.step.p21.StepParser.RealContext;
@@ -20,7 +25,6 @@ import de.bitub.step.p21.StepParser.StringContext;
 import de.bitub.step.p21.StepParser.UntypedContext;
 import de.bitub.step.p21.mapper.NameToClassifierMap;
 import de.bitub.step.p21.util.Antlr4Util;
-import de.bitub.step.p21.util.XPressModel;
 
 public class P21EntityListener extends P21LevelListener
 {
