+++ de.bitub.step.p21/src/de/bitub/step/p21/persistence/P21LoadImpl.java
@@ -38,8 +38,8 @@ import org.eclipse.emf.ecore.util.EcoreUtil;
 import com.google.inject.Guice;
 import com.google.inject.Injector;
 
-import de.bitub.step.p21.P21EntityListener;
 import de.bitub.step.p21.P21Index;
+import de.bitub.step.p21.XPressModel;
 import de.bitub.step.p21.P21IndexImpl.IdStructuralFeaturePair;
 import de.bitub.step.p21.P21IndexImpl.ListTriple;
 import de.bitub.step.p21.concurrrent.P21DataLineRunnable;
@@ -48,8 +48,8 @@ import de.bitub.step.p21.mapper.NameToClassifierMap;
 import de.bitub.step.p21.mapper.NameToClassifierMapImpl;
 import de.bitub.step.p21.mapper.NameToContainerListsMap;
 import de.bitub.step.p21.mapper.NameToContainerListsMapImpl;
+import de.bitub.step.p21.parser.P21EntityListener;
 import de.bitub.step.p21.util.LoggerHelper;
-import de.bitub.step.p21.util.XPressModel;
 
 /**
  * <!-- begin-user-doc -->
